playground/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
psycopg2-binary>=2.9.0
redis>=5.0.0
